+++ Changelog.md
@@ -4,6 +4,7 @@
 * Some monsters now appear in packs (e.g. osquips and rothes)
 * Some monsters produce sounds when attacking
 * Add some missing monster resistances
+* Shatter potions when exposed to fire
 * Consistently call the currency "gold"
 * Consistently show monastery messages as windows
 * Sort stats in obituary matching to order on main screen